requirements.txt
langchain==0.0.354
fastapi==0.108.0
uvicorn==0.25.0
trafilatura==1.6.3
duckduckgo_search==4.1.1
urlextract==1.8.0
qdrant-client==1.7.0
boto3==1.34.12
python-dotenv==1.0.0
pypdf==3.17.4
requests==2.31.0
openai==1.6.1
faiss-cpu==1.7.4
tiktoken==0.5.1